backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
pandas==2.2.1
pyotp==2.9.0
python-dotenv==1.0.1
robin_stocks==3.0.6
scikit_learn==1.4.0
yfinance==0.2.36
